Create fichcomp_IC_2025.lua
This murgaLua script fixes sejour error in "RIMP FICHCOMP isolement contention" file. Sometimes,  an "isolement mesure" is written on a bad sejour which limits (beginning and end) do not include the date of the "mesure". Correcting in base is not allowed (medico-legal aspect), but correcting in "IC FICHCOMP" is possible. This scripts has two files as input : an HQL query export, and the original "FICHCOMP" exported from information system. The output is the corrected FICHCOMP, at same location. MurgaLua has a GUI for searching files (Lua not).